Add tree command
